refactor: styles in separate file
